[FLINK-40150][tests] De-parameterize rescale-history-enabled RescaleTimelineITCase tests
The terminated-reason tests only assert when rescale history is enabled, yet
ran as @TestTemplate: on the disabled parameter each paid for a MiniCluster
startup before assumeThat aborted, and three rebuilt the cluster on top. Move
them to a non-parameterized RescaleTimelineHistoryEnabledITCase that starts one
cluster per test with the configuration it needs. Shared setup and helpers move
to RescaleTimelineITCaseBase; RescaleTimelineITCase keeps the parameterized cases.

Generated-by: Claude Code (Claude Opus 4.8)
